+++ src/entities/concert/api/getConcertCulture.ts
@@ -1,16 +1,16 @@
 import { ApiResponse } from "../../../shared/types/response";
 import axiosInstance from "../../../shared/api/axiosInstance";
 
-type ConcertCulture = {
+export type ConcertCulture = {
   id: number;
   concertId: number;
   content: string;
-  imgUrl: string;
+  title: string;
 };
 
 export async function getConcertCulture(id: number): Promise<ConcertCulture[]> {
   const response = await axiosInstance.get<ApiResponse<ConcertCulture[]>>(
-    `/api/v1/concerts/${id}/cultures`
+    `/api/v2/concerts/${id}/cultures`
   );
   return response.data.data;
 }
