+++ src/pages/ConcertInsidePage.tsx
@@ -2,7 +2,6 @@ import { useEffect } from "react";
 import { useLocation, useParams } from "react-router-dom";
 import ListHeader from "../shared/ui/ListHeader";
 import ConcertInsideInfo from "../entities/concert/ui/ConcertInsideInfo";
-import ConcertCulture from "../widgets/ConcertCulture";
 import PastSetList from "../widgets/PastSetList";
 import ExpectationSetList from "../widgets/ExpectationSetList";
 import { ConcertFilter } from "../entities/concert/types";
