--- src/widgets/ConcertCulture.tsx
@@ -1,17 +1,0 @@
-import { useParams } from "react-router-dom";
-import ConcertCultureCarousel from "../entities/concert/ui/ConcertCultureCarousel";
-
-function ConcertCulture() {
-  const { concertId } = useParams<{ concertId: string }>();
-
-  return (
-    <div className="w-full ml-16 pr-32">
-      <p className="text-grayScaleWhite text-body-lg font-semibold font-NotoSansKR mt-30 mb-14">
-        해당 공연 문화
-      </p>
-      <ConcertCultureCarousel concertId={Number(concertId)} />
-    </div>
-  );
-}
-
-export default ConcertCulture;
